pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vla-serving"
version = "0.1.0"
description = "GPU-agnostic multi-model serving platform for 7B VLA models with ACT-style action chunking"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "Pillow>=10.2.0",
    "av>=12.0.0",
    "pynvml>=11.5.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[project.scripts]
vla-serve = "server.main:main"

[tool.setuptools.packages.find]
include = ["server*", "client*", "common*", "proto*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
